begin rework of protocollib related patches
