+++ src/main/java/me/moomoo/anarchyexploitfixes/AnarchyExploitFixes.java
@@ -4,6 +4,7 @@ import me.moomoo.anarchyexploitfixes.commands.AnarchyExploitFixesCommand;
 import me.moomoo.anarchyexploitfixes.config.Config;
 import me.moomoo.anarchyexploitfixes.config.LanguageCache;
 import me.moomoo.anarchyexploitfixes.modules.AnarchyExploitFixesModule;
+import me.moomoo.anarchyexploitfixes.protocollib.AEFProtocolLibModule;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
@@ -29,6 +30,7 @@ public class AnarchyExploitFixes extends JavaPlugin {
     private static HashMap<String, LanguageCache> languageCacheMap;
     private static int minorMCVersion = Integer.MIN_VALUE;
     private static double tps;
+    private static boolean protocolLib_is_installed;
     public final SortedMap<String, Boolean> enabledModules = new TreeMap<>();
     public final HashSet<UUID> connectionMsgs = new HashSet<>();
     public final HashSet<UUID> playersInNewChunks = new HashSet<>();
@@ -38,6 +40,7 @@ public class AnarchyExploitFixes extends JavaPlugin {
     public void onEnable() {
         instance = this;
         logger = getLogger();
+        protocolLib_is_installed = getServer().getPluginManager().isPluginEnabled("ProtocolLib");
 
         // Fancy enable
         logger.info("                                                          ");
@@ -115,10 +118,10 @@ public class AnarchyExploitFixes extends JavaPlugin {
         reloadLang();
 
         logger.info("Loading Config");
-        reloadAEFConfig();
+        reloadConfiguration();
 
         logger.info("Registering Commands");
-        reloadCommands();
+        AnarchyExploitFixesCommand.reloadCommands();
 
         logger.info("Loading metrics");
         new Metrics(this, 8700);
@@ -146,24 +149,36 @@ public class AnarchyExploitFixes extends JavaPlugin {
 
     public static LanguageCache getLang(CommandSender commandSender) {
         if (commandSender instanceof Player) {
-            Player player = (Player) commandSender;
-            return getLang(player.getLocale());
+            return getLang(((Player) commandSender).getLocale());
         } else {
             return getLang(config.default_lang);
         }
     }
 
-    public void reloadCommands() {
+    public void reloadPlugin() {
+        reloadLang();
+        reloadConfiguration();
         AnarchyExploitFixesCommand.reloadCommands();
     }
 
-    public void reloadAEFConfig() {
+    private void reloadConfiguration() {
         config = new Config();
         AnarchyExploitFixesModule.reloadModules();
+        if (config.protocolLib_IsDisabled) {
+            logger.warning("Not patching NoCom coordinate exploit, BoatFly, PacketFly or CraftingRecipe lag exploit,");
+            logger.warning("because you choose to disable ProtocolLib usage in config.yml");
+        } else {
+            if (protocolLib_is_installed) {
+                AEFProtocolLibModule.reloadModules();
+            } else {
+                logger.severe("Could not find ProtocolLib. Many game breaking exploits can not be patched without it.");
+                logger.severe("Download dependency at https://www.spigotmc.org/resources/protocollib.1997/ or disable in config (not recommended)");
+            }
+        }
         config.saveConfig();
     }
 
-    public void reloadLang() {
+    private void reloadLang() {
         languageCacheMap = new HashMap<>();
         try {
             File langDirectory = new File(getDataFolder() + File.separator + "lang");
