+++ src/main/java/me/moomoo/anarchyexploitfixes/commands/aef/subcommands/ReloadSubCmd.java
@@ -24,10 +24,7 @@ public class ReloadSubCmd extends SubCommand {
     public void perform(CommandSender sender, String[] args) {
         if (sender.hasPermission("anarchyexploitfixes.cmd.reload")) {
             sender.sendMessage(ChatColor.WHITE + "Reloading AnarchyExploitFixes...");
-            AnarchyExploitFixes plugin = AnarchyExploitFixes.getInstance();
-            plugin.reloadLang();
-            plugin.reloadAEFConfig();
-            plugin.reloadCommands();
+            AnarchyExploitFixes.getInstance().reloadPlugin();
             sender.sendMessage(ChatColor.GREEN + "Reload complete.");
         } else {
             sender.sendMessage(ChatColor.translateAlternateColorCodes('&', AnarchyExploitFixes.getLang(sender).noPermission));
