+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/AnarchyExploitFixesModule.java
@@ -1,6 +1,5 @@
 package me.moomoo.anarchyexploitfixes.modules;
 
-import com.comphenix.protocol.ProtocolLibrary;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.modules.bedrock.*;
 import me.moomoo.anarchyexploitfixes.modules.chat.*;
@@ -8,23 +7,33 @@ import me.moomoo.anarchyexploitfixes.modules.chat.commandwhitelist.CommandWhitel
 import me.moomoo.anarchyexploitfixes.modules.chunklimits.*;
 import me.moomoo.anarchyexploitfixes.modules.dupepreventions.*;
 import me.moomoo.anarchyexploitfixes.modules.elytra.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.blocks.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.items.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.items.bannedblocks.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.items.datavalues.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.items.enchantments.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.items.playerheads.*;
-import me.moomoo.anarchyexploitfixes.modules.illegals.items.spawneggs.*;
+import me.moomoo.anarchyexploitfixes.modules.illegals.blocks.PeriodicallyRemoveIllegalBlocks;
+import me.moomoo.anarchyexploitfixes.modules.illegals.blocks.RemoveIllegalBlocksOnChunkload;
+import me.moomoo.anarchyexploitfixes.modules.illegals.blocks.RemoveUnnaturalSpawners;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.RemoveSpecificItemNames;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.RevertOverstacked;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.RevertUnbreakables;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.bannedblocks.BanSpecificBlocks;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.bannedblocks.PreventPlacingBannedBlocks;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.datavalues.IllegalGoldenApples;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.datavalues.RemoveSpecificDataValues;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.enchantments.HigherEnchants;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.enchantments.InapplicableEnchants;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.enchantments.IncompatibleEnchants;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.enchantments.SpecificHigherEnchants;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.playerheads.BanPlayerHeads;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.playerheads.PreventPlacingPlayerHeads;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.spawneggs.PreventUsingSpawnEggs;
+import me.moomoo.anarchyexploitfixes.modules.illegals.items.spawneggs.RemoveSpawnEggs;
 import me.moomoo.anarchyexploitfixes.modules.lagpreventions.*;
 import me.moomoo.anarchyexploitfixes.modules.misc.*;
 import me.moomoo.anarchyexploitfixes.modules.options1b1t.*;
 import me.moomoo.anarchyexploitfixes.modules.patches.*;
-import me.moomoo.anarchyexploitfixes.modules.patches.crashexploits.*;
-import me.moomoo.anarchyexploitfixes.modules.patches.lagexploits.*;
-import me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.boatfly.BoatFlyPatch;
-import me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.craftingrecipelag.CraftingRecipeLag;
-import me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom.NoComPatch;
-import me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.packetfly.PacketFly;
+import me.moomoo.anarchyexploitfixes.modules.patches.crashexploits.DispenserCrash;
+import me.moomoo.anarchyexploitfixes.modules.patches.crashexploits.EndGatewayCrash;
+import me.moomoo.anarchyexploitfixes.modules.patches.crashexploits.MultipleEnderdragons;
+import me.moomoo.anarchyexploitfixes.modules.patches.lagexploits.LeverSpam;
+import me.moomoo.anarchyexploitfixes.modules.patches.lagexploits.SnowballExploit;
 import me.moomoo.anarchyexploitfixes.modules.preventions.*;
 import me.moomoo.anarchyexploitfixes.modules.preventions.netherportals.*;
 import me.moomoo.anarchyexploitfixes.modules.preventions.withers.*;
@@ -47,14 +56,6 @@ public interface AnarchyExploitFixesModule {
         plugin.enabledModules.clear();
         plugin.getServer().getScheduler().cancelTasks(plugin);
         HandlerList.unregisterAll(plugin);
-        if (plugin.getServer().getPluginManager().isPluginEnabled("ProtocolLib")) {
-            ProtocolLibrary.getProtocolManager().removePacketListeners(plugin);
-        } else {
-            if (!AnarchyExploitFixes.getConfiguration().protocolLib_IsDisabled) {
-                AnarchyExploitFixes.getLog().severe("Could not find ProtocolLib. Many game breaking exploits can not be patched without it.");
-                AnarchyExploitFixes.getLog().severe("Download dependency at https://www.spigotmc.org/resources/protocollib.1997/ or disable in config (not recommended)");
-            }
-        }
 
         /*
             Bedrock
@@ -169,11 +170,6 @@ public interface AnarchyExploitFixesModule {
         // Lag Exploits
         modules.add(new SnowballExploit());
         modules.add(new LeverSpam());
-        // Packet Patches
-        modules.add(new BoatFlyPatch());
-        modules.add(new NoComPatch());
-        modules.add(new PacketFly());
-        modules.add(new CraftingRecipeLag());
         /*
             Preventions
         */
