+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/chat/PreventCommandSpam.java
@@ -37,7 +37,7 @@ public class PreventCommandSpam implements AnarchyExploitFixesModule, Listener {
         this.antiSpamCommandTime = config.getInt("chat.prevent-spam.anti-spam-commands.anti-spam-command-time", 1) * 20L;
         this.usingWhitelist = config.getBoolean("chat.prevent-spam.anti-spam-commands.use-whitelist", true, "Exclude commands from antispam.");
         this.whitelistedCommands.addAll(config.getList("chat.prevent-spam.anti-spam-commands.whitelisted-commands", Arrays.asList(
-                "msg", "message", "w", "whisper", "tell", "r", "reply", "/f"
+                "msg", "message", "w", "whisper", "tell", "r", "reply", "f"
         )));
         this.bannedSubCommands.addAll(config.getList("chat.prevent-spam.anti-spam-commands.banned-subcommands", Arrays.asList(
                 "msg Notch", "f perms"
