+++ src/main/java/me/moomoo/anarchyexploitfixes/modules/elytra/ElytraPacketFly.java
@@ -82,7 +82,7 @@ public class ElytraPacketFly implements AnarchyExploitFixesModule, Listener {
                         // drop elytra at player position
                         player.getWorld().dropItemNaturally(player.getLocation(), chestplate);
                         // play sound to notify
-                        player.playSound(player.getLocation(), notifyDing, 1.0F, 1.0F);
+                        player.playSound(player.getEyeLocation(), notifyDing, 1.0F, 1.0F);
                         if (playerShouldBeNotified) player.sendMessage(ChatColor.translateAlternateColorCodes('&',
                                     AnarchyExploitFixes.getLang(player.getLocale()).disablePacketElytraFly
                         ));
