--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/boatfly/BoatFlyPacketListener.java
@@ -1,66 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.boatfly;
-
-import com.comphenix.protocol.PacketType;
-import com.comphenix.protocol.ProtocolLibrary;
-import com.comphenix.protocol.events.ListenerPriority;
-import com.comphenix.protocol.events.PacketAdapter;
-import com.comphenix.protocol.events.PacketEvent;
-import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.utils.LogUtils;
-import org.bukkit.entity.Boat;
-import org.bukkit.entity.Entity;
-import org.bukkit.entity.Player;
-import org.bukkit.scheduler.BukkitScheduler;
-
-import java.util.HashMap;
-import java.util.UUID;
-import java.util.logging.Level;
-
-public class BoatFlyPacketListener {
-    private final BukkitScheduler bukkitScheduler;
-    private final HashMap<UUID, Integer> boatLevels = new HashMap<>();
-    protected BoatFlyPacketListener(
-            final int maxEntityPacketsPer10s,
-            final boolean logIsEnabled,
-            final boolean shouldKickPlayer,
-            final String logPrefix
-    ) {
-        AnarchyExploitFixes plugin = AnarchyExploitFixes.getInstance();
-        this.bukkitScheduler = plugin.getServer().getScheduler();
-        ProtocolLibrary.getProtocolManager().addPacketListener(new PacketAdapter(
-                plugin, ListenerPriority.HIGHEST, PacketType.Play.Client.USE_ENTITY
-        ) {
-            @Override
-            public void onPacketReceiving(PacketEvent event) {
-                if (event.isPlayerTemporary()) return;
-                Player player = event.getPlayer();
-                Entity vehicle = player.getVehicle();
-                if (!(vehicle instanceof Boat)) return;
-
-                UUID playerUniqueID = player.getUniqueId();
-                Integer playerBoatLevels = boatLevels.get(playerUniqueID);
-
-                if (playerBoatLevels != null) {
-                    if (playerBoatLevels > maxEntityPacketsPer10s) {
-                        bukkitScheduler.runTask(plugin, vehicle::remove);
-                        if (shouldKickPlayer) player.kickPlayer(AnarchyExploitFixes.getLang(player.getLocale()).masked_KickMessage);
-                        if (logIsEnabled) LogUtils.moduleLog(Level.INFO, logPrefix,
-                                "Prevented player '"+player.getName()+"' from boat flying"
-                        );
-                    } else {
-                        boatLevels.merge(playerUniqueID, 1, Integer::sum);
-                        bukkitScheduler.runTaskLaterAsynchronously(
-                                plugin, () -> boatLevels.put(playerUniqueID, boatLevels.get(playerUniqueID) - 1), 200L
-                        );
-                    }
-                } else {
-                    boatLevels.put(playerUniqueID, 1);
-                    bukkitScheduler.runTaskLaterAsynchronously(
-                            plugin, () -> boatLevels.put(playerUniqueID, boatLevels.get(playerUniqueID) - 1), 200L
-                    );
-                }
-            }
-        });
-    }
-
-}
