--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/nocom/NoComPacketListener.java
@@ -1,73 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom;
-
-import com.comphenix.protocol.PacketType;
-import com.comphenix.protocol.ProtocolLibrary;
-import com.comphenix.protocol.events.ListenerPriority;
-import com.comphenix.protocol.events.PacketAdapter;
-import com.comphenix.protocol.events.PacketContainer;
-import com.comphenix.protocol.events.PacketEvent;
-import com.comphenix.protocol.wrappers.BlockPosition;
-import com.comphenix.protocol.wrappers.EnumWrappers;
-import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom.packetwrapper.WrapperPlayClientBlockDig;
-import me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom.packetwrapper.WrapperPlayClientUseItem;
-
-public class NoComPacketListener {
-    protected NoComPacketListener(
-            final int noComMaxDistance
-    ) {
-        /*
-                    Code by John200410 & AlexProgrammerDE
-                https://github.com/oldfagorg/CoordinateExploitFix
-        */
-
-        AnarchyExploitFixes plugin = AnarchyExploitFixes.getInstance();
-        ProtocolLibrary.getProtocolManager().addPacketListener(new PacketAdapter(
-                plugin, ListenerPriority.NORMAL, PacketType.Play.Client.BLOCK_DIG
-        ) {
-            @Override
-            public void onPacketReceiving(PacketEvent event) {
-                if(event.getPlayer() == null || event.isPlayerTemporary()) return;
-
-                final PacketContainer packet = event.getPacket();
-                final WrapperPlayClientBlockDig wrappedPacket = new WrapperPlayClientBlockDig(packet);
-                final EnumWrappers.PlayerDigType type = wrappedPacket.getStatus();
-
-                if(
-                        type.equals(EnumWrappers.PlayerDigType.START_DESTROY_BLOCK)
-                        || type.equals(EnumWrappers.PlayerDigType.ABORT_DESTROY_BLOCK)
-                        || type.equals(EnumWrappers.PlayerDigType.STOP_DESTROY_BLOCK)
-                ) {
-                    final BlockPosition blockPos = wrappedPacket.getLocation();
-                    final double distance = event.getPlayer().getLocation().distance(blockPos.toLocation(event.getPlayer().getWorld()));
-
-                    if(distance > noComMaxDistance) {
-                        event.setCancelled(true);
-                    }
-                }
-            }
-        });
-
-        // Bypass - Causing errors on 1.12.2+ - Please let me know if there is a way to fix this. https://paste.ee/p/BdWAH
-        if (AnarchyExploitFixes.getMCVersion() <= 12) {
-            ProtocolLibrary.getProtocolManager().addPacketListener(new PacketAdapter(
-                    plugin, ListenerPriority.NORMAL, PacketType.Play.Client.USE_ITEM
-            ) {
-                @Override
-                public void onPacketReceiving(PacketEvent event) {
-                    if (event.getPlayer() == null || event.isPlayerTemporary()) return;
-
-                    final PacketContainer packet = event.getPacket();
-                    final WrapperPlayClientUseItem wrappedPacket = new WrapperPlayClientUseItem(packet);
-                    final BlockPosition blockPos = wrappedPacket.getLocation();
-                    final double distance = event.getPlayer().getLocation().distance(blockPos.toLocation(event.getPlayer().getWorld()));
-
-                    if (distance > noComMaxDistance) {
-                        event.setCancelled(true);
-                    }
-                }
-            });
-        }
-    }
-
-}
