--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/nocom/NoComPatch.java
@@ -1,60 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom;
-
-import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.utils.LogUtils;
-import me.moomoo.anarchyexploitfixes.modules.AnarchyExploitFixesModule;
-
-import java.util.logging.Level;
-
-public class NoComPatch implements AnarchyExploitFixesModule {
-
-    private final int noComMaxDistance;
-
-    public NoComPatch() {
-        shouldEnable();
-        Config config = AnarchyExploitFixes.getConfiguration();
-        config.addComment("patches.prevent-nocom-coordinate-exploit.enable",  "Prevent the NoCom Coordinate exploit.");
-        this.noComMaxDistance = config.getInt("patches.prevent-nocom-coordinate-exploit.nocom-max-distance", 64);
-    }
-
-    @Override
-    public String name() {
-        return "prevent-nocom-coordinate-exploit";
-    }
-
-    @Override
-    public String category() {
-        return "patches";
-    }
-
-    @Override
-    public void enable() {
-        if (AnarchyExploitFixes.getConfiguration().protocolLib_IsDisabled) {
-            LogUtils.moduleLog(Level.WARNING, name(), "Not enabling because ProtocolLib is disabled in config.");
-        } else {
-            if (AnarchyExploitFixes.getInstance().getServer().getPluginManager().isPluginEnabled("ProtocolLib")) {
-                new NoComPacketListener(noComMaxDistance);
-            } else {
-                LogUtils.moduleLog(Level.SEVERE, name(), "Exploit can't be patched because ProtocolLib is missing.");
-            }
-        }
-    }
-
-    @Override
-    public boolean shouldEnable() {
-        if (AnarchyExploitFixes.getConfiguration().getBoolean("patches.prevent-nocom-coordinate-exploit.enable", true)) {
-            int mcVersion = AnarchyExploitFixes.getMCVersion();
-            if (mcVersion == 17) {
-                if (AnarchyExploitFixes.getInstance().getServer().getVersion().contains("1.17.1")) {
-                    return false;
-                } else {
-                    LogUtils.moduleLog(Level.WARNING, name(), "Consider upgrading to 1.17.1 as NoCom is patched on that version.");
-                }
-            }
-            return mcVersion <= 17;
-        } else {
-            return false;
-        }
-    }
-}
