--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/nocom/packetwrapper/AbstractPacket.java
@@ -1,30 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom.packetwrapper;
-
-import com.comphenix.protocol.PacketType;
-import com.comphenix.protocol.events.PacketContainer;
-import com.google.common.base.Objects;
-
-// Code by John200410 (https://github.com/oldfagorg/CoordinateExploitFix)
-
-public abstract class AbstractPacket {
-	// The packet we will be modifying
-
-	protected PacketContainer handle;
-	
-	/**
-	 * Constructs a new strongly typed wrapper for the given packet.
-	 *
-	 * @param handle - handle to the raw packet data.
-	 * @param type - the packet type.
-	 */
-	protected AbstractPacket(PacketContainer handle, PacketType type) {
-		// Make sure we're given a valid packet
-		if (handle == null)
-			throw new IllegalArgumentException("Packet handle cannot be NULL.");
-		if (!Objects.equal(handle.getType(), type))
-			throw new IllegalArgumentException(handle.getHandle()
-					+ " is not a packet of type " + type);
-		
-		this.handle = handle;
-	}
-}
