--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/nocom/packetwrapper/WrapperPlayClientBlockDig.java
@@ -1,53 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom.packetwrapper;
-
-import com.comphenix.protocol.PacketType;
-import com.comphenix.protocol.events.PacketContainer;
-import com.comphenix.protocol.wrappers.BlockPosition;
-import com.comphenix.protocol.wrappers.EnumWrappers.PlayerDigType;
-
-// Code by John200410 (https://github.com/oldfagorg/CoordinateExploitFix)
-
-public class WrapperPlayClientBlockDig extends AbstractPacket {
-
-	public static final PacketType TYPE = PacketType.Play.Client.BLOCK_DIG;
-	
-	public WrapperPlayClientBlockDig() {
-		super(new PacketContainer(TYPE), TYPE);
-		handle.getModifier().writeDefaults();
-	}
-	
-	public WrapperPlayClientBlockDig(PacketContainer packet) {
-		super(packet, TYPE);
-	}
-	
-	/**
-	 * Retrieve Location.
-	 * <p>
-	 * Notes: block position
-	 *
-	 * @return The current Location
-	 */
-	public BlockPosition getLocation() {
-		return handle.getBlockPositionModifier().read(0);
-	}
-	
-	/**
-	 * Set Location.
-	 *
-	 * @param value - new value.
-	 */
-	public void setLocation(BlockPosition value) {
-		handle.getBlockPositionModifier().write(0, value);
-	}
-	
-	/**
-	 * Retrieve Status.
-	 * <p>
-	 * Notes: the action the player is taking against the block (see below)
-	 *
-	 * @return The current Status
-	 */
-	public PlayerDigType getStatus() {
-		return handle.getPlayerDigTypes().read(0);
-	}
-}
