--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/nocom/packetwrapper/WrapperPlayClientUseItem.java
@@ -1,41 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.nocom.packetwrapper;
-
-import com.comphenix.protocol.PacketType;
-import com.comphenix.protocol.events.PacketContainer;
-import com.comphenix.protocol.wrappers.BlockPosition;
-
-// Code by John200410 (https://github.com/oldfagorg/CoordinateExploitFix)
-
-public class WrapperPlayClientUseItem extends AbstractPacket {
-
-	public static final PacketType TYPE = PacketType.Play.Client.USE_ITEM;
-
-	public WrapperPlayClientUseItem() {
-		super(new PacketContainer(TYPE), TYPE);
-		handle.getModifier().writeDefaults();
-	}
-
-	public WrapperPlayClientUseItem(PacketContainer packet) {
-		super(packet, TYPE);
-	}
-	
-	/**
-	 * Retrieve Location.
-	 * <p>
-	 * Notes: block position
-	 *
-	 * @return The current Location
-	 */
-	public BlockPosition getLocation() {
-		return handle.getBlockPositionModifier().read(0);
-	}
-	
-	/**
-	 * Set Location.
-	 *
-	 * @param value - new value.
-	 */
-	public void setLocation(BlockPosition value) {
-		handle.getBlockPositionModifier().write(0, value);
-	}
-}
