--- src/main/java/me/moomoo/anarchyexploitfixes/modules/patches/packetpatches/packetfly/PacketFly.java
@@ -1,53 +1,0 @@
-package me.moomoo.anarchyexploitfixes.modules.patches.packetpatches.packetfly;
-
-import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
-import me.moomoo.anarchyexploitfixes.config.Config;
-import me.moomoo.anarchyexploitfixes.utils.LogUtils;
-import me.moomoo.anarchyexploitfixes.modules.AnarchyExploitFixesModule;
-
-import java.util.logging.Level;
-
-public class PacketFly implements AnarchyExploitFixesModule {
-
-    private final AnarchyExploitFixes plugin;
-    private final boolean logIsEnabled, shouldKickPlayer;
-    private final int maxTeleportPacketsPer10s;
-
-    public PacketFly() {
-        shouldEnable();
-        this.plugin = AnarchyExploitFixes.getInstance();
-        Config config = AnarchyExploitFixes.getConfiguration();
-        config.addComment("patches.packetfly-patch.enable", "Patches Futureclient / Rusherhack packet fly exploit.");
-        this.logIsEnabled = config.getBoolean("patches.packetfly-patch.log", true);
-        this.maxTeleportPacketsPer10s = config.getInt("patches.packetfly-patch.max-teleport-packets-per-10s", 25);
-        this.shouldKickPlayer = config.getBoolean("patches.packetfly-patch.kick-player-if-packetflying", false);
-    }
-
-    @Override
-    public String name() {
-        return "packetfly-patch";
-    }
-
-    @Override
-    public String category() {
-        return "patches";
-    }
-
-    @Override
-    public void enable() {
-        if (AnarchyExploitFixes.getConfiguration().protocolLib_IsDisabled) {
-            LogUtils.moduleLog(Level.WARNING, name(), "Not enabling because ProtocolLib is disabled in config.");
-        } else {
-            if (AnarchyExploitFixes.getInstance().getServer().getPluginManager().isPluginEnabled("ProtocolLib")) {
-                new PacketFlyListener(plugin, maxTeleportPacketsPer10s, shouldKickPlayer, logIsEnabled, name());
-            } else {
-                LogUtils.moduleLog(Level.SEVERE, name(), "Exploit can't be patched because ProtocolLib is missing.");
-            }
-        }
-    }
-
-    @Override
-    public boolean shouldEnable() {
-        return AnarchyExploitFixes.getConfiguration().getBoolean("patches.packetfly-patch.enable", true);
-    }
-}
