+++ src/main/java/me/moomoo/anarchyexploitfixes/protocollib/elytra/ElytraGlobal.java
@@ -1,39 +1,48 @@
-package me.moomoo.anarchyexploitfixes.modules.elytra;
+package me.moomoo.anarchyexploitfixes.protocollib.elytra;
 
+import com.comphenix.protocol.PacketType;
+import com.comphenix.protocol.ProtocolLibrary;
+import com.comphenix.protocol.events.ListenerPriority;
+import com.comphenix.protocol.events.PacketAdapter;
+import com.comphenix.protocol.events.PacketEvent;
 import com.cryptomorin.xseries.XMaterial;
 import com.cryptomorin.xseries.XSound;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.config.Config;
 import me.moomoo.anarchyexploitfixes.config.LanguageCache;
-import me.moomoo.anarchyexploitfixes.modules.AnarchyExploitFixesModule;
+import me.moomoo.anarchyexploitfixes.protocollib.AEFProtocolLibModule;
 import org.bukkit.ChatColor;
+import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
 
+import java.util.HashMap;
+import java.util.UUID;
+
 import static me.moomoo.anarchyexploitfixes.utils.ElytraUtils.*;
 
-public class ElytraGlobal implements AnarchyExploitFixesModule, Listener {
+public class ElytraGlobal extends PacketAdapter implements AEFProtocolLibModule {
 
     private final AnarchyExploitFixes plugin;
     private static final double tolerance = 0.02;
     private static final Sound notifyDing = XSound.ENTITY_EXPERIENCE_ORB_PICKUP.parseSound();
     private final int spawn_Radius;
-    private final double global_SpeedOldChunks, global_SpeedNewChunks, global_BurstSpeedOldChunks,
-            global_BurstSpeedNewChunks, global_BurstOldChunk_TPS, global_BurstNewChunk_TPS, global_DenyElytraTPS;
+    private final double global_SpeedOldChunks, global_SpeedNewChunks, global_BurstSpeedOldChunks, global_BurstSpeedNewChunks,
+            global_BurstOldChunk_TPS, global_BurstNewChunk_TPS, global_DenyElytraTPS;
     private final boolean showActionbarMsgs, displayChunkInfo, spawn_SettingsEnabled, ceiling_SettingsEnabled, global_DenyElytra,
             global_EnableBursting, global_DenyOnLowTPS, global_AlsoRemoveOnLowTPS;
     private final Material elytra = XMaterial.ELYTRA.parseMaterial();
 
-    public ElytraGlobal() {
+    private final HashMap<UUID, Location> previousLocation = new HashMap<>();
+
+    public ElytraGlobal(AnarchyExploitFixes plugin) {
+        super(plugin, ListenerPriority.HIGHEST, PacketType.Play.Client.POSITION);
+        this.plugin = plugin;
         shouldEnable();
-        this.plugin = AnarchyExploitFixes.getInstance();
+
         Config config = AnarchyExploitFixes.getConfiguration();
         this.showActionbarMsgs = config.getBoolean("elytra.elytra-speed.display-actionbar", true, "Display info in Actionbar while flying.");
         this.displayChunkInfo = config.getBoolean("elytra.elytra-speed.display-chunk-info-in-actionbar", true, "Inform flying player if they are in old or new chunks.");
@@ -67,7 +76,7 @@ public class ElytraGlobal implements AnarchyExploitFixesModule, Listener {
 
     @Override
     public void enable() {
-        plugin.getServer().getPluginManager().registerEvents(this, plugin);
+        ProtocolLibrary.getProtocolManager().addPacketListener(this);
     }
 
     @Override
@@ -75,13 +84,21 @@ public class ElytraGlobal implements AnarchyExploitFixesModule, Listener {
         return AnarchyExploitFixes.getConfiguration().getBoolean("elytra.elytra-speed.Global-Settings.enable", true);
     }
 
-    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
-    private void onPlayerMove(PlayerMoveEvent event) {
+    @Override
+    public void onPacketReceiving(PacketEvent event) {
+        if (event.isPlayerTemporary()) return;
         Player player = event.getPlayer();
-        if (!player.isGliding()) return;
+        if (player == null || !player.isGliding() || player.isInsideVehicle()) return;
         if (player.hasPermission("anarchyexploitfixes.bypass")) return;
         if (spawn_SettingsEnabled && isAtSpawn(player, spawn_Radius)) return;
         if (ceiling_SettingsEnabled && isOnNetherCeiling(player)) return;
+
+        UUID playerUniqueId = player.getUniqueId();
+
+        if (!previousLocation.containsKey(playerUniqueId)) {
+            previousLocation.put(playerUniqueId, player.getLocation());
+            return;
+        }
 
         LanguageCache lang = AnarchyExploitFixes.getLang(player.getLocale());
 
@@ -114,9 +131,10 @@ public class ElytraGlobal implements AnarchyExploitFixesModule, Listener {
             return;
         }
 
-        double flySpeed = blocksPerTick(event.getFrom(), event.getTo());
+        double flySpeed = blocksPerTick(previousLocation.get(playerUniqueId), player.getLocation());
+        previousLocation.put(playerUniqueId, player.getLocation());
 
-        if (plugin.playersInNewChunks.contains(player.getUniqueId())) {
+        if (plugin.playersInNewChunks.contains(playerUniqueId)) {
             // Speed New Chunks
             if (global_EnableBursting && canBurst(global_BurstNewChunk_TPS)) {
                 // Burst Speed New Chunks
@@ -264,6 +282,4 @@ public class ElytraGlobal implements AnarchyExploitFixesModule, Listener {
             }
         }
     }
-
-
 }
