+++ src/main/java/me/moomoo/anarchyexploitfixes/protocollib/elytra/ElytraOnCeiling.java
@@ -1,25 +1,30 @@
-package me.moomoo.anarchyexploitfixes.modules.elytra;
+package me.moomoo.anarchyexploitfixes.protocollib.elytra;
 
+import com.comphenix.protocol.PacketType;
+import com.comphenix.protocol.ProtocolLibrary;
+import com.comphenix.protocol.events.ListenerPriority;
+import com.comphenix.protocol.events.PacketAdapter;
+import com.comphenix.protocol.events.PacketEvent;
 import com.cryptomorin.xseries.XMaterial;
 import com.cryptomorin.xseries.XSound;
 import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
 import me.moomoo.anarchyexploitfixes.config.Config;
 import me.moomoo.anarchyexploitfixes.config.LanguageCache;
-import me.moomoo.anarchyexploitfixes.modules.AnarchyExploitFixesModule;
+import me.moomoo.anarchyexploitfixes.protocollib.AEFProtocolLibModule;
 import org.bukkit.ChatColor;
+import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
 
+import java.util.HashMap;
+import java.util.UUID;
+
 import static me.moomoo.anarchyexploitfixes.utils.ElytraUtils.*;
 
-public class ElytraOnCeiling implements AnarchyExploitFixesModule, Listener {
+public class ElytraOnCeiling extends PacketAdapter implements AEFProtocolLibModule {
 
     private final AnarchyExploitFixes plugin;
     private static final double tolerance = 0.02;
@@ -31,9 +36,13 @@ public class ElytraOnCeiling implements AnarchyExploitFixesModule, Listener {
             ceiling_EnableBursting, ceiling_DenyOnLowTPS, ceiling_AlsoRemoveOnLowTPS;
     private final Material elytra = XMaterial.ELYTRA.parseMaterial();
 
-    public ElytraOnCeiling() {
+    private final HashMap<UUID, Location> previousLocation = new HashMap<>();
+
+    public ElytraOnCeiling(AnarchyExploitFixes plugin) {
+        super(plugin, ListenerPriority.HIGHEST, PacketType.Play.Client.POSITION);
+        this.plugin = plugin;
         shouldEnable();
-        this.plugin = AnarchyExploitFixes.getInstance();
+
         Config config = AnarchyExploitFixes.getConfiguration();
         this.showActionbarMsgs = config.getBoolean("elytra.elytra-speed.display-actionbar", true);
         this.displayChunkInfo = config.getBoolean("elytra.elytra-speed.display-chunk-info-in-actionbar", true);
@@ -65,7 +74,7 @@ public class ElytraOnCeiling implements AnarchyExploitFixesModule, Listener {
 
     @Override
     public void enable() {
-        plugin.getServer().getPluginManager().registerEvents(this, plugin);
+        ProtocolLibrary.getProtocolManager().addPacketListener(this);
     }
 
     @Override
@@ -73,13 +82,21 @@ public class ElytraOnCeiling implements AnarchyExploitFixesModule, Listener {
         return AnarchyExploitFixes.getConfiguration().getBoolean("elytra.elytra-speed.Nether-Ceiling.enable", true);
     }
 
-    @EventHandler(priority = EventPriority.NORMAL, ignoreCancelled = true)
-    private void onPlayerMove(PlayerMoveEvent event) {
+    @Override
+    public void onPacketReceiving(PacketEvent event) {
+        if (event.isPlayerTemporary()) return;
         Player player = event.getPlayer();
-        if (!player.isGliding()) return;
+        if (player == null || !player.isGliding() || player.isInsideVehicle()) return;
         if (player.hasPermission("anarchyexploitfixes.bypass")) return;
         if (!isOnNetherCeiling(player)) return;
         if (spawn_SettingsEnabled && isAtSpawn(player, spawn_Radius)) return;
+
+        UUID playerUniqueId = player.getUniqueId();
+
+        if (!previousLocation.containsKey(playerUniqueId)) {
+            previousLocation.put(playerUniqueId, player.getLocation());
+            return;
+        }
 
         LanguageCache lang = AnarchyExploitFixes.getLang(player.getLocale());
 
@@ -112,9 +129,10 @@ public class ElytraOnCeiling implements AnarchyExploitFixesModule, Listener {
             return;
         }
 
-        double flySpeed = blocksPerTick(event.getFrom(), event.getTo());
+        double flySpeed = blocksPerTick(previousLocation.get(playerUniqueId), player.getLocation());
+        previousLocation.put(playerUniqueId, player.getLocation());
 
-        if (plugin.playersInNewChunks.contains(player.getUniqueId())) {
+        if (plugin.playersInNewChunks.contains(playerUniqueId)) {
             // Speed New Chunks
             if (ceiling_EnableBursting && canBurst(ceiling_BurstNewChunk_TPS)) {
                 // Burst Speed New Chunks
