+++ src/main/java/me/moomoo/anarchyexploitfixes/protocollib/nocom/NoComBlockDigPacket.java
@@ -1,0 +1,86 @@
+package me.moomoo.anarchyexploitfixes.protocollib.nocom;
+
+import com.comphenix.protocol.PacketType;
+import com.comphenix.protocol.ProtocolLibrary;
+import com.comphenix.protocol.events.ListenerPriority;
+import com.comphenix.protocol.events.PacketAdapter;
+import com.comphenix.protocol.events.PacketContainer;
+import com.comphenix.protocol.events.PacketEvent;
+import com.comphenix.protocol.wrappers.BlockPosition;
+import com.comphenix.protocol.wrappers.EnumWrappers;
+import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
+import me.moomoo.anarchyexploitfixes.config.Config;
+import me.moomoo.anarchyexploitfixes.protocollib.AEFProtocolLibModule;
+import me.moomoo.anarchyexploitfixes.utils.LogUtils;
+import org.bukkit.entity.Player;
+
+import java.util.logging.Level;
+
+public class NoComBlockDigPacket extends PacketAdapter implements AEFProtocolLibModule {
+
+    private final int noComMaxDistance;
+
+    public NoComBlockDigPacket(AnarchyExploitFixes plugin) {
+        super(plugin, ListenerPriority.NORMAL, PacketType.Play.Client.BLOCK_DIG);
+        shouldEnable();
+
+        Config config = AnarchyExploitFixes.getConfiguration();
+        config.addComment("patches.prevent-nocom-coordinate-exploit.enable",  "Prevent the NoCom Coordinate exploit.");
+        this.noComMaxDistance = config.getInt("patches.prevent-nocom-coordinate-exploit.nocom-max-distance", 64);
+    }
+
+    @Override
+    public String name() {
+        return "prevent-nocom-coordinate-exploit";
+    }
+
+    @Override
+    public String category() {
+        return "patches";
+    }
+
+    @Override
+    public void enable() {
+        ProtocolLibrary.getProtocolManager().addPacketListener(this);
+    }
+
+    @Override
+    public boolean shouldEnable() {
+        if (AnarchyExploitFixes.getConfiguration().getBoolean("patches.prevent-nocom-coordinate-exploit.enable", true)) {
+            int mcVersion = AnarchyExploitFixes.getMCVersion();
+            if (mcVersion == 17) {
+                if (AnarchyExploitFixes.getInstance().getServer().getVersion().contains("1.17.1")) {
+                    return false;
+                } else {
+                    LogUtils.moduleLog(Level.WARNING, name(), "Consider upgrading to 1.17.1 as NoCom is natively patched on that version.");
+                }
+            }
+            return mcVersion <= 17;
+        } else {
+            return false;
+        }
+    }
+
+    @Override
+    public void onPacketReceiving(PacketEvent event) {
+        if (event.isPlayerTemporary()) return;
+        Player player = event.getPlayer();
+        if (player == null) return;
+
+        final PacketContainer packet = event.getPacket();
+        final EnumWrappers.PlayerDigType type = packet.getPlayerDigTypes().read(0);
+
+        if(
+                type.equals(EnumWrappers.PlayerDigType.START_DESTROY_BLOCK)
+                || type.equals(EnumWrappers.PlayerDigType.ABORT_DESTROY_BLOCK)
+                || type.equals(EnumWrappers.PlayerDigType.STOP_DESTROY_BLOCK)
+        ) {
+            final BlockPosition blockPos = packet.getBlockPositionModifier().read(0);
+
+            if (player.getLocation().distance(blockPos.toLocation(player.getWorld())) > noComMaxDistance) {
+                event.setCancelled(true);
+            }
+        }
+    }
+
+}
