+++ src/main/java/me/moomoo/anarchyexploitfixes/protocollib/nocom/NoComUseItemPacket.java
@@ -1,0 +1,49 @@
+package me.moomoo.anarchyexploitfixes.protocollib.nocom;
+
+import com.comphenix.protocol.PacketType;
+import com.comphenix.protocol.ProtocolLibrary;
+import com.comphenix.protocol.events.ListenerPriority;
+import com.comphenix.protocol.events.PacketAdapter;
+import com.comphenix.protocol.events.PacketEvent;
+import com.comphenix.protocol.wrappers.BlockPosition;
+import me.moomoo.anarchyexploitfixes.AnarchyExploitFixes;
+import me.moomoo.anarchyexploitfixes.protocollib.AEFProtocolLibModule;
+import org.bukkit.entity.Player;
+
+public class NoComUseItemPacket extends PacketAdapter implements AEFProtocolLibModule {
+    private final int noComMaxDistance;
+    public NoComUseItemPacket(AnarchyExploitFixes plugin) {
+        super(plugin, ListenerPriority.NORMAL, PacketType.Play.Client.USE_ITEM);
+        shouldEnable();
+        this.noComMaxDistance = AnarchyExploitFixes.getConfiguration().getInt("patches.prevent-nocom-coordinate-exploit.nocom-max-distance", 64);
+    }
+    @Override
+    public String name() {
+        return "prevent-nocom-coordinate-exploit";
+    }
+    @Override
+    public String category() {
+        return "patches";
+    }
+    @Override
+    public void enable() {
+        ProtocolLibrary.getProtocolManager().addPacketListener(this);
+    }
+    @Override
+    public boolean shouldEnable() {
+        return AnarchyExploitFixes.getConfiguration().getBoolean("patches.prevent-nocom-coordinate-exploit.enable", true);
+    }
+    @Override
+    public void onPacketReceiving(PacketEvent event) {
+        if (event.isPlayerTemporary()) return;
+        Player player = event.getPlayer();
+        if (player == null) return;
+
+        final BlockPosition blockPos = event.getPacket().getBlockPositionModifier().read(0);
+
+        if (player.getLocation().distance(blockPos.toLocation(player.getWorld())) > noComMaxDistance) {
+            event.setCancelled(true);
+        }
+    }
+
+}
